Using explicit EcrImage in cdk if possible
This commit uses regex to test if the image uri matches an ecr image and
if so, it explicitly creates an EcrImage cdk resource.

CDK automatically will handle granting fine-grained permissions when an
EcrImage is used.

This way, we don't need to grant any new permissions explicitly -- cdk
will handle this for us